+++ translator/src/main/scala/soda/translator/io/DirectoryProcessor.scala
@@ -8,19 +8,19 @@ trait DirectoryProcessor {
 
   def process_soda_file: File => Boolean
 
-  lazy val Soda_suffix = ".soda"
+  lazy val soda_suffix = ".soda"
 
   lazy val all_files = DirectoryScannerImpl () .get_all_files (new File (start )  )
 
   lazy val soda_files =
     all_files
       .filter (x => x.isFile )
-      .filter (x => x.getName.endsWith (Soda_suffix )  )
+      .filter (x => x.getName.endsWith (soda_suffix )  )
 
   lazy val lib_files =
     all_files
       .filter (x => x.isFile )
-      .filter (file => file.getName == LibraryDeployerImpl () .Library_marker_file )
+      .filter (file => file.getName == LibraryDeployerImpl () .library_marker_file )
 
   def process (): Boolean =
     LibraryDeployerImpl () .expand_library (lib_files ) &&