+++ translator/src/main/scala/soda/translator/replacement/CharType.scala
@@ -11,13 +11,13 @@ case class CharType (ordinal: Int, name: String ) extends soda.lib.EnumConstant
  */
 trait CharTypeConstant {
 
-  lazy val UndefinedType = CharType (0, "UndefinedType")
-  lazy val QuotesType = CharType (1, "QuotesType")
-  lazy val ApostropheType = CharType (2, "ApostropheType")
-  lazy val BackslashType = CharType (3, "BackslashType")
-  lazy val PlainType = CharType (4, "PlainType")
+  lazy val undefined_type = CharType (0, "undefined_type")
+  lazy val quotes_type = CharType (1, "quotes_type")
+  lazy val apostrophe_type = CharType (2, "apostrophe_type")
+  lazy val backslash_type = CharType (3, "backslash_type")
+  lazy val plain_type = CharType (4, "plain_type")
 
-  lazy val CharTypeValues = Seq (UndefinedType, QuotesType, ApostropheType, BackslashType, PlainType )
+  lazy val CharTypeValues = Seq (undefined_type, quotes_type, apostrophe_type, backslash_type, plain_type )
 }
 
 /**
@@ -27,36 +27,36 @@ trait CharTypeFunction extends CharTypeConstant {
 
   lazy val values = CharTypeValues
 
-  lazy val BackslashChar = '\\'
-  lazy val QuotesChar = '\"'
-  lazy val ApostropheChar = '\''
-  lazy val UnderscoreChar = '_'
-  lazy val SymbolChars = Seq ('!', '#', '$', '%', '&', '(', ')', '*', '+', ',', '-', '.', '/', ':', ';', '<', '=', '>', '?', '@', '[', ']', '^', '`', '{', '|', '}', '~')
+  lazy val backslash_char = '\\'
+  lazy val quotes_char = '\"'
+  lazy val apostrophe_char = '\''
+  lazy val underscore_char = '_'
+  lazy val symbol_chars = Seq ('!', '#', '$', '%', '&', '(', ')', '*', '+', ',', '-', '.', '/', ':', ';', '<', '=', '>', '?', '@', '[', ']', '^', '`', '{', '|', '}', '~')
 
   def get_char_type (ch: Char ): CharType =
-    if (is_quotes (ch ) ) QuotesType
-    else if (is_apostrophe (ch ) ) ApostropheType
-    else if (is_backslash (ch ) ) BackslashType
-    else if (is_whitespace (ch ) || is_letter_or_digit_or_underscore (ch ) || is_symbol (ch ) ) PlainType
-    else UndefinedType
+    if (is_quotes (ch ) ) quotes_type
+    else if (is_apostrophe (ch ) ) apostrophe_type
+    else if (is_backslash (ch ) ) backslash_type
+    else if (is_whitespace (ch ) || is_letter_or_digit_or_underscore (ch ) || is_symbol (ch ) ) plain_type
+    else undefined_type
 
   def is_whitespace (ch: Char ): Boolean =
     ch.isWhitespace
 
   def is_letter_or_digit_or_underscore (ch: Char ): Boolean =
-    ch.isLetterOrDigit || ch == UnderscoreChar
+    ch.isLetterOrDigit || ch == underscore_char
 
   def is_symbol (ch: Char ): Boolean =
-    SymbolChars.contains (ch )
+    symbol_chars.contains (ch )
 
   def is_quotes (ch: Char ): Boolean =
-    ch == QuotesChar
+    ch == quotes_char
 
   def is_apostrophe (ch: Char ): Boolean =
-    ch == ApostropheChar
+    ch == apostrophe_char
 
   def is_backslash (ch: Char ): Boolean =
-    ch == BackslashChar
+    ch == backslash_char
 }
 
 case class CharTypeEnum () extends CharTypeFunction