+++ translator/src/main/scala/soda/translator/replacement/ParserTransition.scala
@@ -8,14 +8,14 @@ case class ParserState (ordinal: Int, name: String ) extends soda.lib.EnumConsta
 
 trait ParserStateConstant {
 
-  lazy val UndefinedState = ParserState (0, "UndefinedState")
-  lazy val QuotesState = ParserState (1, "QuotesState")
-  lazy val ApostropheState = ParserState (2, "ApostropheState")
-  lazy val QuotesBackslashState = ParserState (3, "QuotesBackslashState")
-  lazy val ApostropheBackslashState = ParserState (4, "ApostropheBackslashState")
-  lazy val Plain = ParserState (5, "Plain")
+  lazy val undefined_state = ParserState (0, "undefined_state")
+  lazy val quotes_state = ParserState (1, "quotes_state")
+  lazy val apostrophe_state = ParserState (2, "apostrophe_state")
+  lazy val quotes_backslash_state = ParserState (3, "quotes_backslash_state")
+  lazy val apostrophe_backslash_state = ParserState (4, "apostrophe_backslash_state")
+  lazy val plain = ParserState (5, "plain")
 
-  lazy val ParserStateValues = Seq (UndefinedState, QuotesState, ApostropheState, QuotesBackslashState, ApostropheBackslashState, Plain )
+  lazy val parser_state_values = Seq (undefined_state, quotes_state, apostrophe_state, quotes_backslash_state, apostrophe_backslash_state, plain )
 }
 
 /**
@@ -23,14 +23,14 @@ trait ParserStateConstant {
  */
 trait ParserStateFunction extends ParserStateConstant {
 
-  lazy val values = ParserStateValues
+  lazy val values = parser_state_values
 
   def is_same_class (x: ParserState, y: ParserState ): Boolean =
     (x == y ) || is_like (x, y ) || is_like (y, x )
 
   def is_like (x: ParserState, y: ParserState ): Boolean =
-    (x == QuotesState && y == QuotesBackslashState ) ||
-      (x == ApostropheState && y == ApostropheBackslashState )
+    (x == quotes_state && y == quotes_backslash_state ) ||
+      (x == apostrophe_state && y == apostrophe_backslash_state )
 }
 
 case class ParserStateEnum () extends ParserStateFunction
@@ -39,11 +39,11 @@ trait ParserTransition extends ParserStateConstant with CharTypeConstant {
 
   lazy val TransitionsThatChangeStates: Map [(ParserState, CharType ), ParserState] =
     Map (/* */
-      ((QuotesState, UndefinedType ), UndefinedState ), ((QuotesState, QuotesType ), Plain ), ((QuotesState, BackslashType ), QuotesBackslashState ), /* */
-      ((ApostropheState, UndefinedType ), UndefinedState ), ((ApostropheState, ApostropheType ), Plain ), ((ApostropheState, BackslashType ), ApostropheBackslashState ), /* */
-      ((QuotesBackslashState, UndefinedType ), UndefinedState ), ((QuotesBackslashState, QuotesType ), QuotesState ), ((QuotesBackslashState, ApostropheType ), QuotesState ), ((QuotesBackslashState, BackslashType ), QuotesState ), ((QuotesBackslashState, PlainType ), QuotesState ), /* */
-      ((ApostropheBackslashState, UndefinedType ), UndefinedState ), ((ApostropheBackslashState, QuotesType ), ApostropheState ), ((ApostropheBackslashState, ApostropheType ), ApostropheState ), ((ApostropheBackslashState, BackslashType ), ApostropheState ), ((ApostropheBackslashState, PlainType ), ApostropheState ), /* */
-      ((Plain, UndefinedType ), UndefinedState ), ((Plain, QuotesType ), QuotesState ), ((Plain, ApostropheType ), ApostropheState ), ((Plain, BackslashType ), Plain ), ((Plain, PlainType ), Plain )    )
+      ((quotes_state, undefined_type ), undefined_state ), ((quotes_state, quotes_type ), plain ), ((quotes_state, backslash_type ), quotes_backslash_state ), /* */
+      ((apostrophe_state, undefined_type ), undefined_state ), ((apostrophe_state, apostrophe_type ), plain ), ((apostrophe_state, backslash_type ), apostrophe_backslash_state ), /* */
+      ((quotes_backslash_state, undefined_type ), undefined_state ), ((quotes_backslash_state, quotes_type ), quotes_state ), ((quotes_backslash_state, apostrophe_type ), quotes_state ), ((quotes_backslash_state, backslash_type ), quotes_state ), ((quotes_backslash_state, plain_type ), quotes_state ), /* */
+      ((apostrophe_backslash_state, undefined_type ), undefined_state ), ((apostrophe_backslash_state, quotes_type ), apostrophe_state ), ((apostrophe_backslash_state, apostrophe_type ), apostrophe_state ), ((apostrophe_backslash_state, backslash_type ), apostrophe_state ), ((apostrophe_backslash_state, plain_type ), apostrophe_state ), /* */
+      ((plain, undefined_type ), undefined_state ), ((plain, quotes_type ), quotes_state ), ((plain, apostrophe_type ), apostrophe_state ), ((plain, backslash_type ), plain ), ((plain, plain_type ), plain )    )
 
   def next_parser_state (parser_state: ParserState, char_type: CharType ): ParserState =
     TransitionsThatChangeStates.getOrElse ((parser_state, char_type ), parser_state )