+++ translator/src/test/scala/soda/translator/replacement/CharTypeSpec.scala
@@ -6,7 +6,7 @@ case class CharTypeSpec () extends org.scalatest.funsuite.AnyFunSuite {
    test ("should recognize quotation marks") {
      lazy val input = '"'
      lazy val obtained = CharTypeEnum () .get_char_type (input )
-     lazy val expected = CharTypeEnum () .QuotesType
+     lazy val expected = CharTypeEnum () .quotes_type
 
      assert (obtained == expected )
    }
@@ -14,7 +14,7 @@ case class CharTypeSpec () extends org.scalatest.funsuite.AnyFunSuite {
    test ("should recognize apostrophes") {
      lazy val input = '\''
      lazy val obtained = CharTypeEnum () .get_char_type (input )
-     lazy val expected = CharTypeEnum () .ApostropheType
+     lazy val expected = CharTypeEnum () .apostrophe_type
 
      assert (obtained == expected )
    }
@@ -22,7 +22,7 @@ case class CharTypeSpec () extends org.scalatest.funsuite.AnyFunSuite {
    test ("should recognize backslash") {
      lazy val input = '\\'
      lazy val obtained = CharTypeEnum () .get_char_type (input )
-     lazy val expected = CharTypeEnum () .BackslashType
+     lazy val expected = CharTypeEnum () .backslash_type
 
      assert (obtained == expected )
    }
@@ -30,14 +30,14 @@ case class CharTypeSpec () extends org.scalatest.funsuite.AnyFunSuite {
    test ("should recognize a simple char") {
      lazy val input = 'a'
      lazy val obtained = CharTypeEnum () .get_char_type (input )
-     lazy val expected = CharTypeEnum () .PlainType
+     lazy val expected = CharTypeEnum () .plain_type
 
      assert (obtained == expected )
    }
 
    test ("should recognize plain text") {
      lazy val inputStr = "This is plain text with symbols. 0123456789 _ . !?"
-     lazy val expected = Seq (CharTypeEnum () .PlainType )
+     lazy val expected = Seq (CharTypeEnum () .plain_type )
      lazy val obtained =
        inputStr
          .map (ch => CharTypeEnum () .get_char_type (ch )  )