+++ translator/src/test/scala/soda/translator/replacement/TokenizerSpec.scala
@@ -5,7 +5,7 @@ case class TokenizerSpec () extends org.scalatest.funsuite.AnyFunSuite {
 
   test ("should tokenize a small example") {
     lazy val input = "    val Constant = \"my text\""
-    lazy val expected = Seq (Token ("    val Constant = ", ParserStateEnum () .Plain, 0 ), Token ("\"my text\"", ParserStateEnum () .QuotesState, 19 ), Token ("", ParserStateEnum () .Plain, 28 )    )
+    lazy val expected = Seq (Token ("    val Constant = ", ParserStateEnum () .plain, 0 ), Token ("\"my text\"", ParserStateEnum () .quotes_state, 19 ), Token ("", ParserStateEnum () .plain, 28 )    )
     lazy val obtained = TokenizerImpl (input ) .get_tokens
 
     assert (obtained == expected )
@@ -13,7 +13,7 @@ case class TokenizerSpec () extends org.scalatest.funsuite.AnyFunSuite {
 
   test ("should tokenize a common tab in a string") {
     lazy val input = "  x = \"abc\tde\""
-    lazy val expected = Seq (Token ("  x = ", ParserStateEnum () .Plain, 0 ), Token ("\"abc\tde\"", ParserStateEnum () .QuotesState, 6 ), Token ("", ParserStateEnum () .Plain, 14 )    )
+    lazy val expected = Seq (Token ("  x = ", ParserStateEnum () .plain, 0 ), Token ("\"abc\tde\"", ParserStateEnum () .quotes_state, 6 ), Token ("", ParserStateEnum () .plain, 14 )    )
     lazy val obtained = TokenizerImpl (input ) .get_tokens
 
     assert (obtained == expected )
@@ -21,7 +21,7 @@ case class TokenizerSpec () extends org.scalatest.funsuite.AnyFunSuite {
 
   test ("should tokenize an escaped tab in a string") {
     lazy val input = "  x = \"abc\\tde\""
-    lazy val expected = Seq (Token ("  x = ", ParserStateEnum () .Plain, 0 ), Token ("\"abc\\tde\"", ParserStateEnum () .QuotesState, 6 ), Token ("", ParserStateEnum () .Plain, 15 )    )
+    lazy val expected = Seq (Token ("  x = ", ParserStateEnum () .plain, 0 ), Token ("\"abc\\tde\"", ParserStateEnum () .quotes_state, 6 ), Token ("", ParserStateEnum () .plain, 15 )    )
     lazy val obtained = TokenizerImpl (input ) .get_tokens
 
     assert (obtained == expected )
@@ -29,7 +29,7 @@ case class TokenizerSpec () extends org.scalatest.funsuite.AnyFunSuite {
 
   test ("should tokenize a single function definition") {
     lazy val input = "def f(x: Int): Int = x"
-    lazy val expected = Seq (Token ("def f(x: Int): Int = x", ParserStateEnum () .Plain, 0 )    )
+    lazy val expected = Seq (Token ("def f(x: Int): Int = x", ParserStateEnum () .plain, 0 )    )
     lazy val obtained = TokenizerImpl (input ) .get_tokens
 
     assert (obtained == expected )
@@ -37,10 +37,9 @@ case class TokenizerSpec () extends org.scalatest.funsuite.AnyFunSuite {
 
   test ("should tokenize a function call") {
     lazy val input = "\tas_digits (5 * number)"
-    lazy val expected = Seq (Token ("\tas_digits (5 * number)", ParserStateEnum () .Plain, 0 )    )
+    lazy val expected = Seq (Token ("\tas_digits (5 * number)", ParserStateEnum () .plain, 0 )    )
     lazy val obtained = TokenizerImpl (input ) .get_tokens
 
     assert (obtained == expected )
   }
-
 }